bugfixes and additions
* Fixed Genetic Algorithm with some nasty bugs
* Added option to highlight Gamma Set in InstancePlotter
* LocalSearch::perturbSolution now counts NODES, not moves
* Added parameters to solverapp and adjusted default values
* Pos is back to being double (due to OpenGL funtions...)

+++ mlp/include/iparser/defines.h
@@ -1,4 +1,4 @@
 #pragma
 using Node = std::size_t;
-using Pos = float;
+using Pos = double;
 using Dist = int;
